refactor: Instana's Span structure.
Refactor the Instana's Span structure to improve performance and prevent
potential issues.

Declouped the BaseSpan, RegisteredSpan, and SDKSpan from the span.py
file to prevent possible circular import errors.

Moved all sequences with kind or types of spans to the kind.py file.

Created the ReadableSpan class to provide read-only access to span
attributes and information. Made the InstanaSpan to multiple inherit
from OpenTelemetry's API Span and ReadableSpan classes.
